Try to find share/gosam-contrib/gosam.conf
With this patch, the program tries also the default
${prefix}/share/gosam-contrib/ path to find gosam.conf, where ${prefix}
is assumed to be the same as in the GoSam installation.

This is needed if a user forgets to copy or link gosam.conf of
gosam-contrib to $HOME/ as described in its README and prevents (not yet
described) configuration efforts on standard multi-user installations.

git-svn-id: svn+ssh://svn.hepforge.org/hepforge/svn/gosam/branches/jfsoden-develop@119 a6e02f90-e1d0-4296-904b-e314d7e4afaa