Better support of the "input file without --template" way.
With this patch, the unofficial and undocumented shortcut

$ edit input.rc
$ gosam.py input.rc

where only process and scheme specific options are set in input.rc,
produces the same output as the documented ways

$ edit input.rc
$ gosam.py --template input.in --merge input.rc
$ gosam.py input.in

or

$ gosam.py --template input.in
$ edit input.in
$ gosam.py input.in

The shortcut is detected now and the same system-specific options are
enabled as with --template are put at the end of the output file.

Because the user cannot see or edit all this system-specific settings
with the shortcut, it is (still) neither recommended nor "official
supported" - but works now as expected.

git-svn-id: svn+ssh://svn.hepforge.org/hepforge/svn/gosam/branches/jfsoden-develop@134 a6e02f90-e1d0-4296-904b-e314d7e4afaa