Subprocess-specific settings in input card
This patch enables subprocess-specific settings in the input card.

The new syntax is

   OPTION[LIST_OF_SUBPROCESSES]=VALUE

LIST_OF_SUBPROCESSES is a comma-separated list of
numbers of subprocess for which the OPTION is set to VALUE. Ranges
and exclusion of ranges with "!" (or "^") are supported.

Examples for valid lists:

0-2              = {0,1,2}
-6,!3-4          = {0,1,2,3,5,6}
1-4,!3,9         = {1,2,4,9}

The counting of the suprocesses is the same as in the BLHA order file
and starts at zero to match to the corresponding p*_ subdirectories.
It does not necessarily match the labels returned in the BLHA contract
file.

Subprocess-specific settings need to be unambiguous and overwrite any
otherwise set value.

git-svn-id: svn+ssh://svn.hepforge.org/hepforge/svn/gosam/branches/jfsoden-develop@763 a6e02f90-e1d0-4296-904b-e314d7e4afaa