Support both QCD and QED coupling orders simultaneously.
The BLHA2 keywords CouplingPower QCD and CouplingPower QED or
AlphasPower and AlphaPower are now fully supported.

BSM couplings are not yet supported via BLHA2.

The new experimental syntax in the gosam input file is:
order=QCD,2,4,QED,1,1

The coupling in which the correction should be
performed should be specified first.

git-svn-id: svn+ssh://svn.hepforge.org/hepforge/svn/gosam/branches/jfsoden-develop@839 a6e02f90-e1d0-4296-904b-e314d7e4afaa