Fix parsing of floats vs integers
Because the integer pattern in ExpressionScanner contained the
start of the float pattern, if integer was attempted first (by
golem.model.scanner.Scanner), a string like '1.2' would be lexed
as an integer 1, a dot, and an integer 2.

Unfortunately Scanner attempts patters in the order of
__dict__.items(), which is unspecified, and has changed since
Python 2. In Python 3.7+ the dictionaries are guaranteed to
iterate in the insertion order, so we expect __dict__.items()
to come in declaration order, but in general it is best not to
count on this.

The fix is twofold:
1) to move 'def float()' up, so that it would be used first on
   Python 3.7+;
2) to add lookahead to the integer pattern, so that it would not
   match floats.