Fix git revision parsing for distribution tarballs
The revision is now determined two ways: from the current git
repository, or from the PKG-INFO file. The latter is automatically
created by './setup.py sdist', and contains the full version
string.

This means the source can now be distributed two ways: together
with the full Git repository, or as an archive created with
'setup.py sdist'.

Note that setup.py is expected to be run from its directory; if
not, things might break weirdly.